feat: User specification NOtInRoom

+++ src/main/java/io/solar/specification/filter/UserFilter.java
@@ -11,5 +11,5 @@ public class UserFilter {
 
     private String login;
     private String title;
-    private Integer notInRoom;
+    private Long notInRoom;
 }
